.gitattributes: sync with files currently in the repo
The `export-ignore` directives in the `.gitattributes` file determine which files are excluded when the repository is packaged for distribution, like on tagging a release.

The file exclusion list was out of date.

There is one open question - I'm not sure whether the `webpack` directory needs to be included or excluded. Anyone got some insights into that ?